+++ src/main/java/edu/cafeteria/Repos/StaffRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import edu.cafeteria.model.Staff;
 
 @Repository
-public interface StaffRepository extends JpaRepository<Staff, Integer> {
+public interface StaffRepository extends JpaRepository<Staff, Long> {
 }
